requirements.txt
torch
torchaudio
transformers
numpy
pandas
scikit-learn
scipy
tqdm
nibabel
pyyaml
matplotlib
seaborn
opencv-python-headless